+++ src/LSDW.Abstractions/Application/Missions/ITrafficking.cs
@@ -1,0 +1,10 @@
+﻿using LSDW.Abstractions.Application.Missions.Base;
+
+namespace LSDW.Abstractions.Application.Missions;
+
+/// <summary>
+/// The trafficking mission interface.
+/// </summary>
+public interface ITrafficking : IMission
+{
+}
