--- src/LSDW.Abstractions/Interfaces/Application/Missions/ITrafficking.cs
@@ -1,10 +1,0 @@
-﻿using LSDW.Abstractions.Interfaces.Application.Missions.Base;
-
-namespace LSDW.Abstractions.Interfaces.Application.Missions;
-
-/// <summary>
-/// The trafficking mission interface.
-/// </summary>
-public interface ITrafficking : IMission
-{
-}
