--- src/LSDW.Abstractions/Interfaces/Presentation/Items/ISwitchItem.cs
@@ -1,10 +1,0 @@
-﻿using LSDW.Abstractions.Interfaces.Presentation.Items.Base;
-
-namespace LSDW.Abstractions.Interfaces.Presentation.Items;
-
-/// <summary>
-/// The switch item interface.
-/// </summary>
-public interface ISwitchItem : IItem
-{
-}
