--- src/LSDW.Abstractions/Interfaces/Presentation/Menus/ISettingsMenu.cs
@@ -1,10 +1,0 @@
-﻿using LSDW.Abstractions.Interfaces.Presentation.Menus.Base;
-
-namespace LSDW.Abstractions.Interfaces.Presentation.Menus;
-
-/// <summary>
-/// The settings menu interface.
-/// </summary>
-public interface ISettingsMenu : IMenu
-{
-}
