+++ src/LSDW.Abstractions/Presentation/Items/ISwitchItem.cs
@@ -1,0 +1,10 @@
+﻿using LSDW.Abstractions.Presentation.Items.Base;
+
+namespace LSDW.Abstractions.Presentation.Items;
+
+/// <summary>
+/// The switch item interface.
+/// </summary>
+public interface ISwitchItem : IItem
+{
+}
