+++ src/LSDW.Abstractions/Presentation/Menus/ISettingsMenu.cs
@@ -1,0 +1,10 @@
+﻿using LSDW.Abstractions.Presentation.Menus.Base;
+
+namespace LSDW.Abstractions.Presentation.Menus;
+
+/// <summary>
+/// The settings menu interface.
+/// </summary>
+public interface ISettingsMenu : IMenu
+{
+}
