+++ src/LSDW.Application/Missions/Trafficking.cs
@@ -1,8 +1,8 @@
 ﻿using GTA;
-using LSDW.Abstractions.Interfaces.Application.Missions;
-using LSDW.Abstractions.Interfaces.Application.Providers;
-using LSDW.Abstractions.Interfaces.Infrastructure.Services;
-using LSDW.Abstractions.Interfaces.Presentation.Menus;
+using LSDW.Abstractions.Application.Missions;
+using LSDW.Abstractions.Application.Providers;
+using LSDW.Abstractions.Infrastructure.Services;
+using LSDW.Abstractions.Presentation.Menus;
 using LSDW.Domain.Enumerators;
 using LSDW.Domain.Interfaces.Actors;
 using LSDW.Domain.Interfaces.Models;
