+++ src/LSDW.Domain/Properties/AssemblyInfo.cs
@@ -16,11 +16,11 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCopyrightAttribute("Copyright © 2023 BoBoBaSs84")]
 [assembly: System.Reflection.AssemblyDescriptionAttribute("The modification is an attempt to bring the \"trade\" game mechanics from Chinatown" +
     " Wars to the fictional state of San Andreas.")]
-[assembly: System.Reflection.AssemblyFileVersionAttribute("2023.6.5.639")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("2023.6.5.639-dev")]
+[assembly: System.Reflection.AssemblyFileVersionAttribute("2023.6.5.675")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("2023.6.5.675-dev")]
 [assembly: System.Reflection.AssemblyProductAttribute("Los.Santos.Dope.Wars")]
 [assembly: System.Reflection.AssemblyTitleAttribute("LSDW.Domain")]
-[assembly: System.Reflection.AssemblyVersionAttribute("2023.6.5.639")]
+[assembly: System.Reflection.AssemblyVersionAttribute("2023.6.5.675")]
 [assembly: System.Reflection.AssemblyMetadataAttribute("RepositoryUrl", "https://github.com/BoBoBaSs84/Los.Santos.Dope.Wars")]
 [assembly: System.Resources.NeutralResourcesLanguageAttribute("en-US")]
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("LSDW.Domain.Tests")]
