+++ src/LSDW.Infrastructure/Services/LoggerService.cs
@@ -1,4 +1,4 @@
-﻿using LSDW.Abstractions.Interfaces.Infrastructure.Services;
+﻿using LSDW.Abstractions.Infrastructure.Services;
 using LSDW.Domain.Classes.Models;
 using System.Runtime.CompilerServices;
 using System.Text;
