+++ src/LSDW.Presentation/Items/SwitchItem.cs
@@ -1,5 +1,5 @@
 ﻿using LemonUI.Menus;
-using LSDW.Abstractions.Interfaces.Presentation.Items;
+using LSDW.Abstractions.Presentation.Items;
 using LSDW.Domain.Enumerators;
 using RESX = LSDW.Presentation.Properties.Resources;
