+++ src/LSDW.Presentation/Menus/SettingsMenu.cs
@@ -1,7 +1,7 @@
 ﻿using LemonUI;
 using LemonUI.Menus;
-using LSDW.Abstractions.Interfaces.Infrastructure.Services;
-using LSDW.Abstractions.Interfaces.Presentation.Menus;
+using LSDW.Abstractions.Infrastructure.Services;
+using LSDW.Abstractions.Presentation.Menus;
 using LSDW.Domain.Interfaces.Services;
 using LSDW.Presentation.Items;
 using RESX = LSDW.Presentation.Properties.Resources;
