+++ tests/LSDW.Application.Tests/Helpers/RealTimeProvider.cs
@@ -1,4 +1,4 @@
-﻿using LSDW.Abstractions.Interfaces.Application.Providers;
+﻿using LSDW.Abstractions.Application.Providers;
 
 namespace LSDW.Application.Tests.Helpers;
