+++ tests/LSDW.Application.Tests/Properties/AssemblyInfo.cs
@@ -12,12 +12,12 @@ using System;
 using System.Reflection;
 
 [assembly: System.Reflection.AssemblyCompanyAttribute("LSDW.Application.Tests")]
-[assembly: System.Reflection.AssemblyConfigurationAttribute("Release")]
-[assembly: System.Reflection.AssemblyFileVersionAttribute("2023.6.5.639")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("2023.6.5.639")]
+[assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
+[assembly: System.Reflection.AssemblyFileVersionAttribute("2023.6.5.675")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("2023.6.5.675-dev")]
 [assembly: System.Reflection.AssemblyProductAttribute("LSDW.Application.Tests")]
 [assembly: System.Reflection.AssemblyTitleAttribute("LSDW.Application.Tests")]
-[assembly: System.Reflection.AssemblyVersionAttribute("2023.6.5.639")]
+[assembly: System.Reflection.AssemblyVersionAttribute("2023.6.5.675")]
 [assembly: System.Resources.NeutralResourcesLanguageAttribute("en-US")]
 
 // Generated by the MSBuild WriteCodeFragment class.
