+++ tests/LSDW.Domain.Tests/Properties/AssemblyInfo.cs
@@ -13,11 +13,11 @@ using System.Reflection;
 
 [assembly: System.Reflection.AssemblyCompanyAttribute("LSDW.Domain.Tests")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
-[assembly: System.Reflection.AssemblyFileVersionAttribute("2023.6.5.639")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("2023.6.5.639-dev")]
+[assembly: System.Reflection.AssemblyFileVersionAttribute("2023.6.5.675")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("2023.6.5.675-dev")]
 [assembly: System.Reflection.AssemblyProductAttribute("LSDW.Domain.Tests")]
 [assembly: System.Reflection.AssemblyTitleAttribute("LSDW.Domain.Tests")]
-[assembly: System.Reflection.AssemblyVersionAttribute("2023.6.5.639")]
+[assembly: System.Reflection.AssemblyVersionAttribute("2023.6.5.675")]
 [assembly: System.Resources.NeutralResourcesLanguageAttribute("en-US")]
 
 // Generated by the MSBuild WriteCodeFragment class.
